[IMP] fs_file: Remove storage code; rename FSFileBytesIO -> FSFileValue
