+++ fs_file/readme/CONTRIBUTORS.rst
@@ -1,1 +1,2 @@
 Laurent Mignon <laurent.mignon@acsone.eu>
+Marie Lejeune <marie.lejeune@acsone.eu>
